chore: Refactor create event info form
