+++ src/modules/organization/createEvent/event-info-form/EventInfoForm.vue
@@ -107,28 +107,17 @@
 </template>
 
 <script lang="ts">
-import {
-  computed,
-  defineComponent,
-  reactive,
-  Ref,
-  ref,
-  watch,
-  toRefs
-} from "vue";
+import { defineComponent, toRefs } from "vue";
 import BaseLabelAndTextInput from "@/commons/UI/BaseLabelAndTextInput.vue";
 import SingleNameSelect from "@/commons/UI/select/SingleNameSelect.vue";
 import RemovableTag from "../removable-tag/RemovableTag.vue";
 import BaseDatePicker from "@/commons/UI/BaseDatePicker.vue";
 import BaseUploadImgButton from "@/commons/UI/BaseUploadImgButton.vue";
 import FormPreviewCards from "../form-preview-cards/FormPreviewCards.vue";
-import { validatePhone, isNumber } from "@/commons/utils/validForm";
-import { set } from "date-fns";
 import { hourNames, hourValues } from "@/commons/constant/hour";
 import { UPDATE_MODEL_VALUE } from "@/commons/constant";
-import { Tag } from "@/apollo/types";
 import { EventInfoForm } from "@/commons/Interfaces";
-import { testTags } from "../testData";
+import useEventInfo from "./useEventInfoForm";
 
 export default defineComponent({
   name: "EventInfoForm",
@@ -150,87 +139,23 @@ export default defineComponent({
   setup(props, context) {
     const { modelValue } = toRefs(props);
 
-    const name = ref(modelValue.value.name);
-    const contact = ref(modelValue.value.contact);
-    const tags: Ref<Tag[]> = ref(modelValue.value.tags);
-    const tagSearch: Ref<Tag> = ref({ id: -1, name: "", events: [] });
-    const description = ref(modelValue.value.description);
-    const attendeeLimit = ref(modelValue.value.attendeeLimit.toString());
-    const registrationDueDate = ref(modelValue.value.registrationDueDate);
-    const registrationDueTime = ref(0);
-    const posterImg = ref(modelValue.value.posterImg);
-    const coverImg = ref(modelValue.value.coverImg);
-
-    //Get from API
-    const tagOptionValues: Tag[] = testTags;
-    const tagOptionNames: string[] = tagOptionValues.map(tag => tag.name);
-
-    const isValidContact = computed(() => {
-      return validatePhone(contact.value);
-    });
-
-    const isValidAttendeeLimit = computed(() => {
-      return isNumber(attendeeLimit.value);
-    });
-
-    const attendeeLimitNumber = computed(() => {
-      if (!isValidAttendeeLimit.value || attendeeLimit.value === "") return 0;
-      return parseInt(attendeeLimit.value);
-    });
-
-    const registrationDueDateTime = computed(() => {
-      const dateTime = set(new Date(registrationDueDate.value), {
-        hours: registrationDueTime.value
-      });
-      return dateTime.toString();
-    });
-
-    const isValidEventInfo = computed(() => {
-      if (
-        name.value === "" ||
-        description.value === "" ||
-        attendeeLimitNumber.value === 0
-      )
-        return false;
-      return true;
-    });
-
-    function addTag(tag: Tag) {
-      const tagId = tag.id;
-      const isInTagsList = tags.value.some(
-        selectedTag => selectedTag.id === tagId
-      );
-      if (isInTagsList) return;
-      tags.value.push(tagSearch.value);
-    }
-
-    function removeTag(index: number) {
-      tags.value.splice(index, 1);
-    }
-
-    watch(tagSearch, () => {
-      if (tagSearch.value.id === -1) return;
-      addTag(tagSearch.value);
-      tagSearch.value = { id: -1, name: "", events: [] };
-    });
-
-    const eventInfo: EventInfoForm = reactive({
-      name: name,
-      contact: contact,
-      tags: tags,
-      description: description,
-      attendeeLimit: attendeeLimitNumber,
-      registrationDueDate: registrationDueDateTime,
-      posterImg: posterImg,
-      coverImg: coverImg,
-      isValid: isValidEventInfo
-    });
-
-    watch(
-      () => eventInfo,
-      () => context.emit(UPDATE_MODEL_VALUE, eventInfo),
-      { deep: true }
-    );
+    const {
+      name,
+      contact,
+      tags,
+      tagSearch,
+      tagOptionNames,
+      tagOptionValues,
+      description,
+      attendeeLimit,
+      registrationDueDate,
+      registrationDueTime,
+      isValidContact,
+      isValidAttendeeLimit,
+      removeTag,
+      posterImg,
+      coverImg
+    } = useEventInfo(modelValue, context);
 
     return {
       name,
@@ -246,12 +171,10 @@ export default defineComponent({
       isValidContact,
       isValidAttendeeLimit,
       removeTag,
-      hourNames,
-      hourValues,
-      registrationDueDateTime,
       posterImg,
       coverImg,
-      eventInfo
+      hourNames,
+      hourValues
     };
   }
 });
